\name{addcaption}
\alias{addcaption}
\title{Add caption}
\description{Add a caption to a referenced object}
\usage{addcaption(key, caption, envir)}
\arguments{
  \item{key}{identifier for the object being referenced}
  \item{caption}{The caption text}
  \item{envir}{The environment where the referenced object resides}
}
\details{This is an internal function called by 
add() and label(). It should not be called directly.}
\value{Caption is updated. No return value.}

\author{Michael Koohafkan}





