\name{addFigure}
\alias{addFigure}
\title{create a figure}
\description{create a reference to a figure (e.g. a ggplot variable) }
\usage{addFigure(key, figure, caption)}
\arguments{
  \item{key}{identifier for the figure being referenced}
  \item{figure}{the figure object, or a path to an image file}
  \item{caption}{the figure caption (optional)}
}

\value{Figure is indexed. No return value}

\author{Michael Koohafkan}





