\name{addTable}
\alias{addTable}
\title{create a table}
\description{create a reference to a table (e.g. a ggplot variable) }
\usage{addTable(key, table, caption)}
\arguments{
  \item{key}{identifier for the table being referenced}
  \item{table}{the table object}
  \item{caption}{the table caption (optional)}
}

\value{Table is indexed. No return value}

\author{Michael Koohafkan}





