\name{anchorChunk}
\alias{anchorChunk}
\title{Anchor chunk}
\description{Produce an anchor tag for a chunk using the chunk name}
\usage{anchorChunk(key)}
\arguments{
  \item{key}{the key for the chunk referenced. It is recommended you use the chunk label}
}

\value{anchor text to be processed by the anchor hook}

\author{Michael Koohafkan}





