\name{anchorFigure}
\alias{anchorFigure}
\title{Anchor for cross-referencing}
\description{Add an HTML anchor tag for figure referencing}
\usage{anchorFigure(key)}
\arguments{
  \item{key}{identifier for the object being referenced}
}

\value{an html anchor tag}

\author{Michael Koohafkan}





