\name{anchorTable}
\alias{anchorTable}
\title{Anchor for cross-referencing}
\description{Add an HTML anchor tag for table referencing}
\usage{anchorTable(key)}
\arguments{
  \item{key}{identifier for the table being referenced}
}

\value{an html anchor tag}

\author{Michael Koohafkan}





