% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{centerCounts}
\alias{centerCounts}
\title{Center single-cell ATAC-seq count data}
\usage{
centerCounts(obj, doInChunks = TRUE, chunkSize = 1000)
}
\arguments{
\item{obj}{Either a \code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}} or \code{\link[Matrix]{dgeMatrix-class}} object of the single-cell ATAC data (peaks x cells) to center raw data for}

\item{doInChunks}{boolean value whether or not to score cells in chunks (useful for large scATAC datasets of > 5,000 cells). If TRUE, cells are centered sequentially in chunks of size=1000 cells at a time to avoid memory limitations, and eventually merged (column-wise). Overriden and set to TRUE if > 10,000 cells in dataset}

\item{chunkSize}{numeric specifying the number of cells to perform centering for at once, if running in chunks (to save memory). Default is 1000}
}
\value{
Either a \code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}} or \code{\link[Matrix]{dgeMatrix-class}} object with counts centered by mean reads in peaks per cell
}
\description{
Function to center raw read counts for scATAC-seq data based on mean reads in features/rows (e.g. peaks) per cell
}
\author{
Vinay Kartha
}
