\name{citeChunk}
\alias{citeChunk}
\title{Cite code chunk}
\description{Cite a code chunk}
\usage{citeChunk(key, link)}
\arguments{
  \item{key}{the chunk label}
  \item{link}{logical: If TRUE, relative link to chunk is included}
}

\value{a chunk citation for cross-referencing.}

\author{Michael Koohafkan}





