\name{citeTable}
\alias{citeTable}
\title{get table number}
\description{return the table reference}
\usage{citeTable(key, link = TRUE, caption = FALSE)}
\arguments{
  \item{key}{identifier for the table being referenced}
  \item{link}{logical: If TRUE, relative link to table is included}
  \item{caption}{Logical: If TRUE, the entire caption is returned.}
}

\value{a table citation for cross-reference or labeling.}

\author{Michael Koohafkan}





