% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DORCs.R
\name{dorcJPlot}
\alias{dorcJPlot}
\title{DORC J scatter plot}
\usage{
dorcJPlot(
  dorcTab,
  cutoff = 7,
  labelTop = 25,
  returnGeneList = FALSE,
  cleanLabels = TRUE,
  labelSize = 4,
  ...
)
}
\arguments{
\item{dorcTab}{data.frame object containing significant peak-gene pairs using which DORC scores will be computed. Must be a filtered set returned from \code{\link[BuenRTools]{runGenePeakcorr}}. IMPORTANT: Make sure the exact same scATAC SE (peak set) was used when determinin DORCs that is used here to get corresponding DORC peak counts}

\item{cutoff}{numeric indicating cut-off to use for number of significant peaks per gene to call it a DORC. Default is 7 significant peaks per gene.}

\item{labelTop}{numeric indicating the number of top genes to add labels for. Setting this to a very high number can lead to plot rendering difficulties (since labels get crowded).}

\item{returnGeneList}{boolean indicating whether to also return the DORCs passing the provided filter. Default is FALSE (will just plot and nothing else)}

\item{cleanLabels}{boolean indicating whether or not to try to have cleaner DORC gene labels. If TRUE, uses \code{\link[ggrepel]{geom_text_repel}}, if FALSE, will use regular \code{\link[ggplot2]{geom_text}}}

\item{labelSize}{numeric indicating what to set label size to. Default is 4.}

\item{...}{additional parameters passed to either \code{\link[ggrepel]{geom_text_repel}} (if cleanLabels is set to TRUE) or \code{\link[ggplot2]{geom_text}}. For example, set size to control font size, or segment.length or segment.color to control call-out segment length and color etc.}
}
\value{
a ggplot object of the DORC J plot. If returnGeneList is set to TRUE, then also returns the vector of DORC genes passing the specified threshold
}
\description{
Function to visualize genes based on number of significant gene-peak correlations (per gene), used to define DORCs
}
\author{
Vinay Kartha
}
