\name{ExternalFigure}
\alias{ExternalFigure}
\title{Place external image}
\description{Place a Figure from a file.}
\usage{ExternalFigure(key)}
\arguments{
  \item{key}{identifier for the figure being referenced}
}

\value{No direct return value. Prints HTML to the console
to be read 'asis' by a knitr code chunk.}

\author{Michael Koohafkan}





