% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractTFNames}
\alias{extractTFNames}
\title{Fetch TF names}
\usage{
extractTFNames(motifIDs)
}
\arguments{
\item{motifIDs}{vector of full motif IDs that are part of the original human or mouse pwms matrix (see \url{https://github.com/GreenleafLab/chromVARmotifs})}
}
\value{
a vector of TF names extracted from the full motif ID (if older format with '_'), or returned as is otherwise
}
\description{
This function fetches TF names from full motif IDs (see \url{https://github.com/GreenleafLab/chromVARmotifs})
}
\author{
Vinay Kartha
}
