% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_convert.R
\name{fig_convert}
\alias{fig_convert}
\title{Convert pdfs to tifs for rendering}
\usage{
fig_convert(dir, silent = TRUE)
}
\arguments{
\item{dir}{Directory to converts}

\item{silent}{If FALSE converted pdf list printed}
}
\value{
Returns a table with column of text links and text labels.
}
\description{
Given a directory will convert all pdf's to tif's to
allow RMarkdown rendering to look pretty. Requires the
R magick package to work.
}
