% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_create.R
\name{fig_create}
\alias{fig_create}
\title{Create a label and link table from a file list or directory}
\usage{
fig_create(filenames = NULL, dir = NULL, desc = FALSE,
  extensions = ".(tif|png)$")
}
\arguments{
\item{filenames}{A character vector of filenames}

\item{dir}{A path to a directory of image files}

\item{desc}{If TRUE table is sorted on descending filename or sort key}

\item{extensions}{Regex to filter file extensions}
}
\value{
Returns a table with column of text links and text labels.
}
\description{
This fucntion is given either a list of filenames or pointed to a directory.
It will creat a table with formatted link and label text for inclusion
in a RMarkdown document. Filenames are filtered to only include those
with specified file extensions.
}
\examples{
\dontrun{
fig_create(dir = ".", desc = TRUE)
}

}
