% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_html2pdf.R
\name{fig_html2pdf}
\alias{fig_html2pdf}
\title{Uses system call to pandoc to convert html to pdf}
\usage{
fig_html2pdf(filename)
}
\arguments{
\item{filename}{Filename of html file}
}
\value{
Returns filename for new file
}
\description{
HTML to PDF conversion via pandoc system call. Writes new
file with old filename but upodates extension to .pdf
}
\examples{
\dontrun{
fig_html2pdf("myfile.html")
}

}
