% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_utils.R
\name{fig_label}
\alias{fig_label}
\title{Create a named label - defaults to 'home'}
\usage{
fig_label(label = "home")
}
\arguments{
\item{label}{The label tag}
}
\value{
Returns nothing, sideeffect is printing label
}
\description{
This function creates an html label tag manually, defaulting
to 'home'. Likely to be used to give a return location from
figures,
}
\examples{
\dontrun{
fig_label()
}

}
