% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_table.R
\name{fig_table}
\alias{fig_table}
\title{Writes out a figr table}
\usage{
fig_table(fig_tbl, show_all = FALSE)
}
\arguments{
\item{fig_tbl}{A figr_table}

\item{show_all}{If TRUE long filename output too}
}
\value{
Returns fig_tbl, sideeffect is to print it
}
\description{
This fucntion will insert a figr table dropping the long filenames
unless these are specifically requested
}
\examples{
\dontrun{
fig_table(tbl)
}

}
