\name{figr-package}
\alias{figr-package}
\alias{figr}
\docType{package}
\title{Simple figure and table indexer for use with document generation.}
\description{This package is designed to be a companion to packages such as knitr. It provides a simple reference manager for visuals generated by R code.}
\details{
\tabular{ll}{Package: \tab figr\cr
Type: \tab Package\cr
Title: \tab Simple figure and table indexer for use with document generation.\cr
Version: \tab 0.1\cr
Date: \tab 2014-05-30\cr
Author: \tab Michael Koohafkan\cr
Maintainer: \tab Michael Koohafkan <michael.koohafkan@gmail.com>\cr
Depends: \tab R (>= 3.1.0)\cr
Suggests: \tab knitr, ggplot2\cr
VignetteBuilder: \tab knitr\cr
License: \tab GPL\cr}
}
\author{Michael Koohafkan}

\keyword{ package }


