\name{getCaption}
\alias{getCaption}
\title{Get caption}
\description{Get the figure or table caption.}
\usage{getCaption(key, envir)}
\arguments{
  \item{key}{identifier for the object being referenced}
  \item{envir}{the environment where the referenced object resides}
}
\details{This is an internal function called by placeFigure() and citeFigure().
It should not be called directly.}
\value{The caption of the specified figure or table.}

\author{Michael Koohafkan}





