% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DORCs.R
\name{getDORCScores}
\alias{getDORCScores}
\title{DORC scATAC-seq score summarization for single cells}
\usage{
getDORCScores(
  ATAC.se,
  dorcTab,
  normalizeATACmat = TRUE,
  geneList = NULL,
  nCores = 4
)
}
\arguments{
\item{ATAC.se}{SummarizedExperiment object of the scATAC-seq reads in peak counts.}

\item{dorcTab}{data.frame object containing significant peak-gene pairs using which DORC scores will be computed. Must be a filtered set returned from \code{\link[FigR]{runGenePeakcorr}}. IMPORTANT: Make sure the exact same scATAC SE (peak set) was used when determining DORCs that is used here to get corresponding DORC peak counts}

\item{normalizeATACmat}{boolean indicating whether or not to normalize the counts present in the ATAC.se object prior to summing peak counts per gene. Default is TRUE (i.e. assumes peak counts are raw).}

\item{geneList}{character vector specifying a subset of genes to compute scores for. Useful if you already have a set list of DORCs and you only want to compute scores for those genes/peaks.}

\item{nCores}{numeric indicating the number of cores to use if parallelizing tasks}
}
\value{
a Matrix object of scores pertaining to the sum of normalized scATAC peak counts per gene, as determined for significant gene-peak pairs when calling DORCs
}
\description{
Function to compute single cell DORC scores for each gene identified as a DORC (see \code{\link[FigR]{runGenePeakcorr}} for determining DORCs through peak-gene correlations)
}
\author{
Vinay Kartha
}
