\name{labelTable}
\alias{labelTable}
\title{Table caption }
\description{Place a table caption}
\usage{labelTable(key, punc = ":")}
\arguments{
  \item{key}{identifier for the table being referenced}
  \item{link}{logical: If TRUE, relative link to figure is included}    \item{punc}{separator to use between reference number and caption
Defaults to ":"}
}

\value{The table caption}

\author{Michael Koohafkan}





