% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{motifPeakZtest}
\alias{motifPeakZtest}
\title{Peak-based TF motif enrichment Z-test}
\usage{
motifPeakZtest(peakSet, bgPeaks, tfMat)
}
\arguments{
\item{peakSet}{vector of peak indices you wish to test for motif enrichment}

\item{bgPeaks}{background peaks matrix returned by \code{\link[chromVAR]{getBackgroundPeaks}} function in chromVAR}

\item{tfMat}{binary match \code{\link[Matrix]{sparseMatrix}} object of (reference) peaks by TF motifs returned by \code{\link[motifmatchr]{matchMotifs}} function}
}
\value{
data frame of motif enrichment results rank sorted by significance and frequency of motifs in the peakset, respectively
}
\description{
Wrapper function to assess a specific set of peaks for enrichment with respect to TF binding motifs using a Z test after adjusting for GC bias and Tn5 accessibility
}
\author{
Vinay Kartha, Samuel Rose, and Jason D. Buenrostro
}
