\name{placeFigure}
\alias{placeFigure}
\title{figure placement and ranking}
\description{places a figure in the document, i.e. determines
the figure number for cross-referencing}
\usage{placeFigure(key)}
\arguments{
  \item{key}{identifier for the object being referenced}
}

\value{the referenced figure}

\author{Michael Koohafkan}





