\name{placeTable}
\alias{placeTable}
\title{table placement and ranking}
\description{places a table in the document, i.e. determines
the table number for cross-referencing}
\usage{placeTable(key)}
\arguments{
  \item{key}{identifier for the table being referenced}
}

\value{the referenced table}

\author{Michael Koohafkan}





