% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{plotDrivers}
\alias{plotDrivers}
\title{Plot FigR scatter profile}
\usage{
plotDrivers(figR.d, marker, score.cut = 1, label = TRUE)
}
\arguments{
\item{figR.d}{data.frame of results returned by \code{\link[FigR]{runFigR}}).}

\item{marker}{Matrix object of smoothed single-cell RNA expression values}

\item{score.cut}{numeric specifying the absolute regulation score to threshold TF-DORC connections on. Default is 1}

\item{myLabels}{character vector specifying the subset of DORCs to test, if not running on everything}
}
\value{
a ggplot2 object of the scatter plot of predicted TF drivers for the specified DORC
}
\description{
Scatter plot visualization of filtered TF-DORC associations based on the enrichment of each motif among the queried DORC's peaks and the correlation of the TF RNA to the DORC accessibility score
}
\author{
Vinay Kartha
}
