% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{plotfigRHeatmap}
\alias{plotfigRHeatmap}
\title{Plot FigR heatmap}
\usage{
plotfigRHeatmap(figR.d, score.cut = 1, DORCs = NULL, TFs = NULL, ...)
}
\arguments{
\item{figR.d}{data.frame of results returned by \code{\link[FigR]{runFigRGRN}}).}

\item{score.cut}{numeric specifying the absolute regulation score to threshold TF-DORC connections on. Default is 1}

\item{DORCs}{character specifying valid DORC gene symbols to subset heatmap to. Default is NULL (no subsetting)}

\item{TFs}{character specifying valid TF gene symbols to subset heatmap to. Default is NULL (no subsetting)}

\item{...}{additional parameters passed to the \code{\link[ComplexHeatmap]{Heatmap}})}
}
\value{
a TF-DORC filtered Heatmap generatd using \code{\link[ComplexHeatmap]{Heatmap}})
}
\description{
Heatmap visualization of TF-DORC associations based on the regulation scores inferred by FigR
}
\author{
Vinay Kartha
}
