% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPairing.R
\name{plotPairs}
\alias{plotPairs}
\title{Plot ATAC-RNA scOptMatch pairs}
\usage{
plotPairs(ATAC, RNA, umap.df, max.show = 300, seed = 123, pairPointSize = 1)
}
\arguments{
\item{ATAC}{character vector of the ATAC barcodes returned in the resulting pairing data frame returned by the \code{\link[FigR]{cell_pairing}} function}

\item{RNA}{character vector of the RNA barcodes returned in the resulting pairing data frame returned by the \code{\link[FigR]{cell_pairing}} function}

\item{umap.df}{data.frame that has the ATAC/RNA co-embedding UMAP 1/2 coordinates in the first and second columns, respectively. Must have valid ATAC and RNA barcodes as row names, which will be used to subset pairs}

\item{max.show}{numeric specifying the maximum number of pairs to limit plotting to. Default is 300 (choose fewer for better visibility if the UMAP is crowded)}

\item{seed}{numeric specifying a random seed to set for sampling pairs (for reproducibility). Default is 123}

\item{pairPointSize}{scatter plot point size. Default is 1}
}
\value{
a ggplot scatter plot with user-provided coordinates, highlighting ATAC-RNA pairs
}
\description{
Function to plot ATAC-RNA pairs in a joint embedding space (.e.g  2D UMAP of both ATAC and RNA cells)
}
\author{
Vinay Kartha
}
