% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{rankDrivers}
\alias{rankDrivers}
\title{Rank TF drivers}
\usage{
rankDrivers(
  figR.d,
  rankBy = c("meanScore", "nTargets"),
  myLabels = NULL,
  score.cut = NULL,
  interactive = FALSE
)
}
\arguments{
\item{figR.d}{data.frame of results returned by \code{\link[FigR]{runFigR}})}

\item{rankBy}{character specifying one of "meanScore" or "nTargets" to either rank TFs by the mean regulation score across all genes, or by the total number of inferred activated or repressed targets passing a specified (absolute) regulation score, respectively}

\item{myLabels}{character vector specifying the subset of DORCs to test, if not running on everything}

\item{score.cut}{numeric specifying the absolute regulation score to threshold TF-DORC connections on, only if rankBy is set to "nTargets". Default is 1 if "nTargets" and no custom cut-off is specified}

\item{interactive}{boolean indicating whether or not to allow interactive hover-over utility for more label information (useful if visualizing too many TFs and labels are hard to distinguish). Default is FALSE}
}
\value{
a ggplot2 object of the resulting plot
}
\description{
Ranked plot of TF activators and repressors based on their inferred regulation score
}
\author{
Vinay Kartha
}
