\name{specenv}
\alias{specenv}
\title{Environment specifier}
\description{specify environment to use.}
\usage{specenv(type)}
\arguments{
  \item{type}{the type of object to be placed, e.g. "Figure" or "Table"}
}
\details{This is an internal function used by figr for 
object management. It should not be called directly.}
\value{The environment to be used}

\author{Michael Koohafkan}





