% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitAndFetch}
\alias{splitAndFetch}
\title{Split and fetch string subparts}
\usage{
splitAndFetch(vec, delim, part)
}
\arguments{
\item{vec}{character vector to apply splitting to}

\item{delim}{character specifying the delimitor to separate string by}

\item{part}{integer specifying the part to fetch after splitting}
}
\value{
substring based on splitting criteria
}
\description{
Function to split and fetch string components based on a specific delimiter
}
\examples{
my_string <- c("This_is_random_string.111","Yes_it_is.222")
# Get first part splitting on _
splitAndFetch(vec=my_string,delim="_",part=1)
# Get first two parts splitting on _
splitAndFetch(vec=my_string,delim="_",part=1:2)
# Get second part splitting on .
splitAndFetch(vec=my_string,delim=".",part=2)
}
\author{
Vinay Kartha
}
