\name{updateRank}
\alias{updateRank}
\title{update rankings of environment}

\usage{updateRank(key, envir)}
\arguments{
  \item{key}{identifier for the object being referenced}
  \item{envir}{the environment being updated}
}
\details{This is an internal function called by placeFigure() and citeFigure().
It should not be called directly.}
\value{Rank of figure or table specified by key.}

\author{Michael Koohafkan}





