\name{addFigure}
\alias{addFigure}
\title{create a figure}
\description{create a reference to a figure (e.g. a plot variable) }
\usage{addFigure(key, figure)}
\arguments{
  \item{key}{the figure label}
  \item{figure}{the plot object}
}

\value{No return value}

\author{Who wrote it}





