\name{citeFigure}
\alias{citeFigure}
\title{get figure number}
\description{return the figure reference}
\usage{citeFigure(key, link = TRUE)}
\arguments{
  \item{key}{identifier for the figure being referenced}
  \item{link}{logical: If TRUE, relative link to figure is included}
}

\value{a figure citation for cross-reference or labeling.}

\author{Michael Koohafkan}





