\name{citeFigure}
\alias{citeFigure}
\title{get figure number}
\description{return the figure reference}
\usage{citeFigure(key, caption = FALSE)}
\arguments{
  \item{key}{identifier for the object being referenced}
  \item{caption}{Logical: If TRUE, the entire caption is returned.}
}

\value{a figure citation for cross-reference or labeling.}

\author{Michael Koohafkan}





