\name{citeTable}
\alias{citeTable}
\title{get table number}
\description{return the table reference}
\usage{citeTable(key, link)}
\arguments{
  \item{key}{identifier for the table being referenced}
  \item{link}{logical: If TRUE, relative link to figure is included}
}

\value{a table citation for cross-reference or labeling.}

\author{Michael Koohafkan}





