% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_figure.R
\name{fig_figure}
\alias{fig_figure}
\title{Insert a figure from a figr table}
\usage{
fig_figure(figr_tbl, index, pagebreak = TRUE)
}
\arguments{
\item{figr_tbl}{A figr table}

\item{pagebreak}{If TRUE inserts pagebreak before figure}

\item{x}{Either an integer index or short filename}
}
\value{
Returns nothing, sideeffect is inserting figure
}
\description{
This function is given a figr table and an index or short filename
and will then insert the corresponding figure into the html document
together with a label to permit linking.
}
\examples{
\dontrun{
fig_figure(tbl, 7)
}

}
