% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_utils.R
\name{fig_link}
\alias{fig_link}
\title{Create a named link - defaults to 'home'}
\usage{
fig_link(label = "home", text = "back")
}
\arguments{
\item{label}{The link tag}

\item{text}{The text to display (deafults to 'back')}
}
\value{
Returns nothing, sideeffect is printing link
}
\description{
This function creates a link tag manually, defaulting
to 'home'. Likely to be used to give a return link from
figures,
}
\examples{
\dontrun{
fig_link()
}

}
