\name{labelFigure}
\alias{labelFigure}
\title{Figure caption}
\description{Place a figure caption}
\usage{labelFigure(key, punc = ":")}
\arguments{
  \item{key}{identifier for the figure being referenced}
  \item{punc}{seperator to use between reference number and caption
Defaults to ":"}
}

\value{The figure caption}

\author{Michael Koohafkan}





