\name{labelTable}
\alias{labelTable}
\title{Table caption }
\description{Place a table caption}
\usage{labelTable(key, punc = ":")}
\arguments{
  \item{key}{identifier for the table being referenced}
  \item{punc}{seperator to use between reference number and caption
Defaults to ":"}
}

\value{The table caption}

\author{Michael Koohafkan}





