% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPairing.R
\name{pairCells}
\alias{pairCells}
\title{Pair cells}
\usage{
pairCells(ATAC, RNA, keepUnique = FALSE, ...)
}
\arguments{
\item{keepUnique}{boolean indicating whether or not to remove any resulting many-to-many ATAC-RNA paired cells returned by the geodesic pairing method (depending on what the `max_multimatch` parameter is set to in {FigR}[cell_pairing], default 5). Default is FALSE, meaning in the returned pairing, there can a single cell in the larger dataset paired to upto `max_multimatch` cells in the smaller dataset. See {FigR}[cell_pairing] for more details. If set to TRUE, we further filter pairs and return for each cell in the larger dataset the best match in the smaller dataset (among the multiple options, if any) based on the min euclidean distances between pair options (in the suoplied reduced dimension space)}

\item{...}{additional parameters passed to {FigR}[cell_pairing]}
}
\description{
Wrapper function for running ATAC/RNA pairing in cell chunks using co-embedding components
}
