\name{placeFigure}
\alias{placeFigure}
\title{figure placement and ranking}
\description{places a figure in the document, i.e. determines
the figure number for cross-referencing}
\usage{placeFigure(key)}
\arguments{
  \item{key}{the figure label}
}
\details{If you place a figure that was already 
placed previously, the figure rank is overwritten.}
\value{the referenced figure object}

\author{Who wrote it}





