% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{plotDrivers}
\alias{plotDrivers}
\title{Plot FigR scatter profile}
\usage{
plotDrivers(figR.d, marker, score.cut = 1, label = TRUE)
}
\arguments{
\item{figR.d}{data.frame of results returned by \code{\link[FigR]{runFigRGRN}})}

\item{marker}{character specifying a valid DORC gene to restrict TF drivers for}

\item{score.cut}{numeric specifying the absolute regulation score to threshold TF-DORC connections on. Default is 1}

\item{label}{boolean indicating whether or not to add text labels for TF drivers passing score filter}
}
\value{
a ggplot2 object of the scatter plot of predicted TF drivers for the specified DORC
}
\description{
Scatter plot visualization of filtered TF-DORC associations based on the enrichment of each motif among the queried DORC's peaks and the correlation of the TF RNA to the DORC accessibility score
}
\author{
Vinay Kartha
}
